remove : room.tsx
